# swagger.yml
openapi: 3.0.3
info:
  title: "NUPEVID API"
  description: "API do sistema NUPEVID para gerenciamento de vítimas, atendimentos e medidas protetivas"
  version: "1.0.0"
  contact:
    name: "Equipe NUPEVID"
servers:
  - url: "/api/v1"
    description: "Servidor Principal"
...
security:
  - bearerAuth: []
  - api_key: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Autenticação via Token JWT obtido no endpoint /auth/login"
    api_key:
      type: apiKey
      in: header
      name: api_key
      description: "Chave de API configurada nas variáveis de ambiente"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Tipo de erro"
          example: "Bad Request"
        message:
          type: string
          description: "Mensagem de erro descritiva"
          example: "Validation error: field X is required"
        status_code:
          type: integer
          format: int32
          description: "Código de status HTTP"
          example: 400
...
    CityComplement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "PORTO VELHO"
        state:
          type: string
          example: "RO"
        battalion:
          type: string
          example: "1ºBPM"
        is_deleted:
          type: boolean
...
    PhoneType:
      type: string
      enum:
        - "Mobile"
        - "Residential"
        - "Work"
      description: "Tipo de telefone"
    ViolenceType:
      type: string
      enum:
        - "Physical"
        - "Sexual"
        - "Psychological"
        - "Moral"
      description: "Tipo de violência sofrida pela vítima"
    HasChildren:
      type: boolean
      description: "Indicador de filhos da vítima"
    SecurityForce:
      type: string
      enum:
        - "Military Police"
        - "Civil Police"
        - "Penal Police"
        - "Fire Department"
        - "Federal Highway Police"
        - "Federal Police"
        - "Municipal Guard"
      description: "Força de segurança do agressor (quando aplicável)"
    RelationshipToVictim:
      type: string
      enum:
        - "Spouse"
        - "Ex-Spouse"
        - "Mother"
        - "Father"
        - "Stepfather"
        - "Stepmother"
        - "Son/Daughter"
        - "Sibling"
        - "Cousin"
        - "Grandfather"
        - "Grandmother"
        - "Brother/Sister-in-law"
        - "Father/Mother-in-law"
        - "Son-in-law"
        - "Daughter-in-law"
        - "Uncle/Aunt"
      description: "Relacionamento do agressor com a vítima"
    EducationLevel:
      type: string
      enum:
        - "Elementary"
        - "High School"
        - "College"
        - "Postgraduate"
        - "Illiterate"
        - "Semi-illiterate"
        - "Master"
        - "Doctorate"
      description: "Nível de escolaridade"
    ProtectiveMeasureStatus:
      type: string
      enum:
        - "Valid"
        - "Revoked"
        - "Expired"
      description: "Status da medida protetiva"
...
    VictimPhone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "(69) 99999-8888"
        phone_type:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PhoneType"
...
    AddressType:
      type: string
      enum:
        - "Residential"
        - "Work"
        - "Correspondence"
        - "Commercial"
        - "Institutional"
        - "Temporary"
        - "Other"
      description: "Tipo de endereco da vitima"
...
    VictimComplement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        cpf:
          type: string
          nullable: true
        birth_date:
          type: string
          format: date
          nullable: true
        city_id:
          type: string
          format: uuid
        education_level:
          allOf:
            - $ref: "#/components/schemas/EducationLevel"
          nullable: true
        occupation:
          type: string
          nullable: true
          example: "Professora"
        has_children:
          $ref: "#/components/schemas/HasChildren"
        children_count:
          type: integer
          nullable: true
          example: 2
        is_pregnant:
          type: boolean
          nullable: true
        has_special_needs:
          type: boolean
        special_needs_type:
          type: array
          items:
            type: string
          nullable: true
        uses_alcohol:
          type: boolean
        uses_drugs:
          type: boolean
        has_psychiatric_issues:
          type: boolean
        psychiatric_issues_type:
          type: array
          items:
            type: string
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/VictimPhone"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/VictimAddress"
        is_deleted:
          type: boolean
    VictimAddress:
      type: object
      required:
        - address_type
      properties:
        id:
          type: string
          format: uuid
        street:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        city_id:
          type: string
          format: uuid
          nullable: true
        zip_code:
          type: string
          nullable: true
        complement:
          type: string
          nullable: true
        address_type:
          $ref: "#/components/schemas/AddressType"
...
    OffenderPhone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "(69) 99999-8888"
        phone_type:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PhoneType"
    OffenderAddress:
      type: object
      required:
        - address_type
      properties:
        id:
          type: string
          format: uuid
        street:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        city_id:
          type: string
          format: uuid
          nullable: true
        zip_code:
          type: string
          nullable: true
        complement:
          type: string
          nullable: true
        address_type:
          $ref: "#/components/schemas/AddressType"
...
    OffenderComplement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        cpf:
          type: string
          nullable: true
        birth_date:
          type: string
          format: date
          nullable: true
        city_id:
          type: string
          format: uuid
        imprisoned:
          type: boolean
        occupation:
          type: string
          nullable: true
        is_public_security_agent:
          type: boolean
        security_force:
          allOf:
            - $ref: "#/components/schemas/SecurityForce"
          nullable: true
        uses_alcohol:
          type: boolean
        uses_drugs:
          type: boolean
        has_psychiatric_issues:
          type: boolean
        psychiatric_issues_type:
          type: array
          items:
            type: string
          nullable: true
        education_level:
          $ref: "#/components/schemas/EducationLevel"
        observation:
          type: string
          nullable: true
        phones:
          type: array
          items:
            $ref: "#/components/schemas/OffenderPhone"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/OffenderAddress"
        is_deleted:
          type: boolean
...
    ProtectiveMeasure:
      type: object
      properties:
        id:
          type: string
          format: uuid
        process_number:
          type: string
        sei_process_number:
          type: string
          nullable: true
          description: "Número do processo no sistema SEI"
        occurrence_report_number:
          type: string
          nullable: true
          description: "Número do boletim de ocorrência"
        issued_at:
          type: string
          format: date
        valid_until:
          type: string
          format: date
          nullable: true
          description: "Data de validade da medida protetiva"
        judicial_authority:
          type: string
        court_district_id:
          type: string
          format: uuid
        distance_meters:
          type: integer
          nullable: true
          description: "Distância em metros que o agressor deve manter da vítima"
        status:
          $ref: "#/components/schemas/ProtectiveMeasureStatus"
        violence_types:
          type: array
          items:
            $ref: "#/components/schemas/ViolenceType"
        relationship_to_victim:
          $ref: "#/components/schemas/RelationshipToVictim"
        assaults_children:
          type: boolean
          description: "Indica se o agressor agride os filhos"
        was_drunk_during_assault:
          type: boolean
        victim_id:
          type: string
          format: uuid
        offender_id:
          type: string
          format: uuid
          description: "ID do agressor associado à medida protetiva"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
    CreateProtectiveMeasureRequest:
      type: object
      required:
        - process_number
        - issued_at
        - judicial_authority
        - court_district_id
        - status
        - violence_types
        - relationship_to_victim
        - assaults_children
        - was_drunk_during_assault
        - victim_id
        - offender_id
      properties:
        process_number:
          type: string
          example: "12345-67.2025.8.26.0000"
        sei_process_number:
          type: string
          nullable: true
          example: "SEI-2025/001234"
        occurrence_report_number:
          type: string
          nullable: true
          example: "BO-2025/001234"
          description: "Número do boletim de ocorrência"
        issued_at:
          type: string
          format: date
        valid_until:
          type: string
          format: date
          nullable: true
        judicial_authority:
          type: string
        court_district_id:
          type: string
          format: uuid
        distance_meters:
          type: integer
          nullable: true
          example: 500
        status:
          $ref: "#/components/schemas/ProtectiveMeasureStatus"
        violence_types:
          type: array
          items:
            $ref: "#/components/schemas/ViolenceType"
        relationship_to_victim:
          $ref: "#/components/schemas/RelationshipToVictim"
        assaults_children:
          type: boolean
        was_drunk_during_assault:
          type: boolean
        victim_id:
          type: string
          format: uuid
        offender_id:
          type: string
          format: uuid
        extensions:
          type: array
          description: "Prorrogações opcionais a serem criadas junto com a medida"
          items:
            $ref: "#/components/schemas/ProtectiveMeasureExtensionUpsert"
    UpdateProtectiveMeasureRequest:
      type: object
      required:
        - process_number
        - issued_at
        - judicial_authority
        - court_district_id
        - status
        - violence_types
        - relationship_to_victim
        - assaults_children
        - was_drunk_during_assault
        - victim_id
        - offender_id
      properties:
        process_number:
          type: string
        sei_process_number:
          type: string
          nullable: true
        occurrence_report_number:
          type: string
          nullable: true
        issued_at:
          type: string
          format: date
        valid_until:
          type: string
          format: date
          nullable: true
        judicial_authority:
          type: string
        court_district_id:
          type: string
          format: uuid
        distance_meters:
          type: integer
          nullable: true
        status:
          $ref: "#/components/schemas/ProtectiveMeasureStatus"
        violence_types:
          type: array
          items:
            $ref: "#/components/schemas/ViolenceType"
        relationship_to_victim:
          $ref: "#/components/schemas/RelationshipToVictim"
        assaults_children:
          type: boolean
        was_drunk_during_assault:
          type: boolean
        victim_id:
          type: string
          format: uuid
        offender_id:
          type: string
          format: uuid
        extensions:
          type: array
          description: "Prorrogações opcionais para criar/atualizar (id opcional)"
          items:
            $ref: "#/components/schemas/ProtectiveMeasureExtensionUpsert"

    # --- Extension Schemas ---
    ProtectiveMeasureExtensionUpsert:
      type: object
      required:
        - extension_number
        - extension_date
      properties:
        id:
          type: string
          format: uuid
          description: "Use para atualizar uma prorrogação existente; omita para criar"
        extension_number:
          type: integer
          example: 1
        extension_date:
          type: string
          format: date
        new_valid_until:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    ProtectiveMeasureExtension:
      type: object
      properties:
        id:
          type: string
          format: uuid
        protective_measure_id:
          type: string
          format: uuid
        extension_number:
          type: integer
          description: "Número da prorrogação (1ª, 2ª, 3ª, etc.)"
          example: 1
        extension_date:
          type: string
          format: date
          description: "Data em que a prorrogação foi concedida"
        new_valid_until:
          type: string
          format: date
          nullable: true
          description: "Nova data de validade após a prorrogação"
        notes:
          type: string
          nullable: true
          description: "Observações sobre a prorrogação"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
...
    ProtectiveMeasureWithExtensions:
      type: object
      allOf:
        - $ref: "#/components/schemas/ProtectiveMeasure"
        - type: object
          properties:
            extensions:
              type: array
              items:
                $ref: "#/components/schemas/ProtectiveMeasureExtension"
    ProtectiveMeasureWithExtensionsAndEntities:
      type: object
      allOf:
        - $ref: "#/components/schemas/ProtectiveMeasure"
        - type: object
          properties:
            extensions:
              type: array
              items:
                $ref: "#/components/schemas/ProtectiveMeasureExtension"
            victim:
              $ref: "#/components/schemas/VictimComplement"
            offender:
              $ref: "#/components/schemas/OffenderComplement"
            court_district:
              $ref: "#/components/schemas/CityComplement"

    # --- Attendance Victim Schemas ---
...
paths:
...
  /protective-measures:
    get:
      tags:
        - "Protective Measures"
      summary: "Listar todas as medidas protetivas"
      description: "Retorna lista de medidas protetivas com filtros baseados nas permissões"
      operationId: "getAllProtectiveMeasures"
      security:
        - bearerAuth: []
          api_key: []
      parameters:
        - name: "page"
          in: "query"
          required: false
          schema:
            type: integer
            example: 1
          description: "Número da página (1-based)"
        - name: "page_size"
          in: "query"
          required: false
          schema:
            type: integer
            example: 10
          description: "Quantidade de itens por página (max 200)"
        - name: "include_complement_for_entities"
          in: "query"
          required: false
          schema:
            type: boolean
            example: false
          description: "Quando true, retorna ProtectiveMeasureWithExtensionsAndEntities (com victim, offender, court_district). Quando false/omitido, retorna ProtectiveMeasureWithExtensions."
      responses:
        "200":
          description: "Lista de medidas protetivas retornada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      oneOf:
                        - $ref: "#/components/schemas/ProtectiveMeasureWithExtensions"
                        - $ref: "#/components/schemas/ProtectiveMeasureWithExtensionsAndEntities"
                  page:
                    type: integer
                    example: 1
                  page_size:
                    type: integer
                    example: 10
                  total_items:
                    type: integer
                    example: 100
                  total_pages:
                    type: integer
                    example: 10
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para listar medidas protetivas"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403

    post:
      tags:
        - "Protective Measures"
      summary: "Criar nova medida protetiva"
      description: "Cria uma nova medida protetiva. Apenas uma medida ativa por vítima é permitida"
      operationId: "createProtectiveMeasure"
      security:
        - bearerAuth: []
          api_key: []
      parameters:
        - name: "include_complement_for_entities"
          in: "query"
          required: false
          schema:
            type: boolean
            example: false
          description: "Quando true, inclui entidades relacionadas completas (victim, offender, court_district) sem created_at/updated_at"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProtectiveMeasureRequest"
      responses:
        "201":
          description: "Medida protetiva criada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  data:
                    oneOf:
                      - $ref: "#/components/schemas/ProtectiveMeasure"
                      - $ref: "#/components/schemas/ProtectiveMeasureWithExtensionsAndEntities"
        "400":
          description: "Dados inválidos ou vítima já possui medida ativa"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Bad Request"
                message: "Bad Request: victim already has an active protective measure"
                status_code: 400
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para criar medidas protetivas"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
  /protective-measures/{id}:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: string
          format: uuid
        description: "ID da medida protetiva"

    get:
      tags:
        - "Protective Measures"
      summary: "Obter medida protetiva por ID"
      description: "Retorna detalhes de uma medida protetiva específica"
      operationId: "getProtectiveMeasureById"
      security:
        - bearerAuth: []
          api_key: []
      parameters:
        - name: "include_complement_for_entities"
          in: "query"
          required: false
          schema:
            type: boolean
            example: false
          description: "Quando true, inclui entidades relacionadas completas (victim, offender, court_district) sem created_at/updated_at"
      responses:
        "200":
          description: "Medida protetiva encontrada"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    oneOf:
                      - $ref: "#/components/schemas/ProtectiveMeasureWithExtensions"
                      - $ref: "#/components/schemas/ProtectiveMeasureWithExtensionsAndEntities"
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para visualizar esta medida protetiva"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Medida protetiva não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: protective measure not found"
                status_code: 404

    put:
      tags:
        - "Protective Measures"
      summary: "Atualizar medida protetiva"
      description: "Atualiza informações de uma medida protetiva existente"
      operationId: "updateProtectiveMeasure"
      security:
        - bearerAuth: []
          api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProtectiveMeasureRequest"
      responses:
        "200":
          description: "Medida protetiva atualizada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/ProtectiveMeasure"
        "400":
          description: "Dados inválidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para atualizar medida protetiva"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Medida protetiva não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: protective measure not found"
                status_code: 404

    delete:
      tags:
        - "Protective Measures"
      summary: "Deletar medida protetiva (soft delete)"
      description: "Realiza soft delete de uma medida protetiva"
      operationId: "deleteProtectiveMeasure"
      security:
        - bearerAuth: []
          api_key: []
      responses:
        "200":
          description: "Medida protetiva deletada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/ProtectiveMeasure"
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para deletar medida protetiva"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Medida protetiva não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: protective measure not found"
                status_code: 404
...